rename: 카드 컴포저블 경로 변경
