+++ app/src/main/java/nextstep/payments/new_card/NewCardScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import nextstep.payments.PaymentCard
+import nextstep.payments.component.PaymentCard
 
 @Composable
 fun NewCardScreen(
